+++ R/benchmarks.R
@@ -16,16 +16,10 @@ benchmarks <- function(run_id = NULL, batch_id = NULL, run_reason = NULL, simpli
 
   if (!is.null(batch_id)) {
     req <- req_url_query(req, batch_id = paste0(batch_id, collapse = ","))
-    if (!utils::packageVersion("httr2") > "0.2.1") {
-      req$url <- curl::curl_unescape(req$url)
-    }
   }
 
   if (!is.null(run_id)) {
     req <- req_url_query(req, run_id = paste0(run_id, collapse = ","))
-    if (!utils::packageVersion("httr2") > "0.2.1") {
-      req$url <- curl::curl_unescape(req$url)
-    }
   }
   resp <- conbench_perform(req)