+++ NEWS.md
@@ -1,6 +1,7 @@
 # conbenchcoms 0.0.2
 
 * Add run_reason argument to `benchmarks`
+* Hard dependency on httr 0.2.2 for escaping capabilities
 
 # conbenchcoms 0.0.1